refactor(storage): move is_expired to utils module
- Rename util.rs to utils.rs for consistency
- Move is_expired function from Storage to utils module
- Remove unused chrono::Utc import from storage.rs
