+++ crates/storage/src/storage_set.rs
@@ -8,7 +8,7 @@ use crate::set::member_key::SetMemberKey;
 use crate::storage::Storage;
 use crate::string::meta::MetaKey;
 use crate::string::meta::SetMetaValue;
-use crate::util::user_key_prefix;
+use crate::utils::user_key_prefix;
 
 impl Storage {
 	pub async fn sadd(&self, key: Bytes, members: Vec<Bytes>) -> Result<u64, StorageError> {
