+++ crates/storage/src/utils.rs
@@ -1,6 +1,12 @@
 use bytes::BufMut;
 use bytes::Bytes;
 use bytes::BytesMut;
+use chrono::Utc;
+
+/// Check if a given expire_ts (milliseconds since epoch) has passed.
+pub fn is_expired(expire_ts: Option<i64>) -> bool {
+	expire_ts.is_some_and(|ts| ts <= Utc::now().timestamp_millis())
+}
 
 /// Build the common storage prefix: len(user_key) (u16 BE) + user_key.
 pub fn user_key_prefix(key: &Bytes) -> Bytes {
